# Reduce memory allocations when computing a tree hash